Further work on update() function
Removal of `__updateIndices()` and `__updateIndirectionChain()` helper functions for now. `__updateIndices()` was redundant with index.py's `update_all_indices()`.

Introduction of `__getLatestTailRid()` to walk through all base and consolidated pages and grab the indirection_rid after that for the tail page.

Only thing left should be to properly update the indirection column of the base page?