Generalized `__getLatestTailRid()` to `__getLatestConRid()`
Since we can get the latest tail_rid from the base page or the latest consolidated page, I generalized it so the helper function returns the consolidated page instead. 

Generalization allows me to update the indirection column of the latest consolidated/base record to point to the updated tail while maintaining the option of getting the latest tail from the newest record. It can be applied to other functions in `query.py` as well.